feat: Integration with signalR and order queue service (rabbitMQ) + order module

+++ Anotaki/Controllers/ExtraController.cs
@@ -1,6 +1,7 @@
 ﻿using anotaki_api.DTOs.Requests.Extra;
 using anotaki_api.DTOs.Response.Api;
 using anotaki_api.Models;
+using anotaki_api.Services;
 using anotaki_api.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -9,15 +10,20 @@ namespace anotaki_api.Controllers
 {
     [Route("api/v1/extra")]
     [ApiController]
-    public class ExtraController(IExtraService extraService)
+    public class ExtraController(IExtraService extraService, IUserService userService) : ControllerBase
     {
         private readonly IExtraService _extraService = extraService;
+        private readonly IUserService _userService = userService;
 
         [HttpGet]
         public async Task<IActionResult> GetAllExtras()
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _extraService.GetAllExtras();
                 return ApiResponse.Create("Extras retrieved successfully.", StatusCodes.Status200OK, data);
             }
@@ -33,6 +39,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _extraService.CreateExtra(dto);
                 return ApiResponse.Create("Extra created successfully.", StatusCodes.Status201Created, data);
             }
@@ -47,6 +57,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 await _extraService.DeleteExtra(extraId);
                 return ApiResponse.Create("Extra delete successfully.", StatusCodes.Status200OK, extraId);
             }
@@ -61,6 +75,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _extraService.UpdateExtra(extra);
                 return ApiResponse.Create("Extra Updated", StatusCodes.Status200OK, data);
             }
@@ -76,6 +94,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _extraService.AddMultipleExtrasToProduct(productId, extraIds);
                 return ApiResponse.Create("Add Multiple Extras to Product", StatusCodes.Status200OK, data);
             } catch (Exception ex)
