+++ Anotaki/Controllers/ProductController.cs
@@ -1,6 +1,7 @@
 ﻿using anotaki_api.DTOs.Requests.Product;
 using anotaki_api.DTOs.Response.Api;
 using anotaki_api.Models;
+using anotaki_api.Services;
 using anotaki_api.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -10,15 +11,20 @@ namespace anotaki_api.Controllers
     [Route("api/v1/product")]
     [Authorize(Roles = Roles.Admin)]
     [ApiController]
-    public class ProductController(IProductService productService)
+    public class ProductController(IProductService productService, IUserService userService) : ControllerBase
     {
         private readonly IProductService _productService = productService;
+        private readonly IUserService _userService = userService;
 
         [HttpPost]
         public async Task<IActionResult> CreateProduct(CreateProductRequestDTO dto)
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _productService.CreateProduct(dto);
                 return ApiResponse.Create("Product Created", StatusCodes.Status201Created, data);
             }
@@ -34,6 +40,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _productService.GetAllProducts();
                 return ApiResponse.Create("Getting All Products", StatusCodes.Status200OK, data);
             }
@@ -48,6 +58,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 var data = await _productService.UpdateProduct(product);
                 return ApiResponse.Create("Product Updated", StatusCodes.Status200OK, data);
             }
@@ -62,6 +76,10 @@ namespace anotaki_api.Controllers
         {
             try
             {
+                var user = await _userService.GetContextUser(User);
+                if (user == null)
+                    return ApiResponse.Create("User not found.", StatusCodes.Status404NotFound);
+
                 await _productService.DeleteProduct(id);
                 return ApiResponse.Create("Product Deleted", StatusCodes.Status200OK, id);
             } catch (Exception ex)
